added a factory bean for currentScreenDimensionSupplier, and inject it into GazeDeviceManagers
