+++ gazeplay-commons/src/main/java/net/gazeplay/commons/gaze/devicemanager/GazeDeviceManagerFactory.java
@@ -1,18 +1,26 @@
 package net.gazeplay.commons.gaze.devicemanager;
 
+import javafx.geometry.Dimension2D;
+import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.commons.configuration.ActiveConfigurationContext;
 import net.gazeplay.commons.configuration.Configuration;
 import net.gazeplay.commons.gaze.EyeTracker;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Supplier;
 
 @Component
 @Slf4j
 public class GazeDeviceManagerFactory {
 
     private final AtomicReference<GazeDeviceManager> currentInstanceReference = new AtomicReference<>();
+
+    @Autowired
+    @Setter
+    private Supplier<Dimension2D> currentScreenDimensionSupplier;
 
     public GazeDeviceManagerFactory() {
     }
@@ -47,7 +55,7 @@ public class GazeDeviceManagerFactory {
             default:
                 gazeDeviceManager = new AbstractGazeDeviceManager() {
                     @Override
-                    public void init() {
+                    public void init(Supplier<Dimension2D> currentScreenDimensionSupplier) {
                     }
 
                     @Override
@@ -58,7 +66,7 @@ public class GazeDeviceManagerFactory {
                 };
         }
 
-        gazeDeviceManager.init();
+        gazeDeviceManager.init(currentScreenDimensionSupplier);
         return gazeDeviceManager;
     }
