+++ gazeplay/src/main/java/net/gazeplay/GazePlay.java
@@ -1,25 +1,17 @@
 package net.gazeplay;
 
 import javafx.beans.property.ReadOnlyBooleanProperty;
-import javafx.collections.ObservableList;
 import javafx.geometry.Dimension2D;
 import javafx.scene.Scene;
-import javafx.stage.Screen;
 import javafx.stage.Stage;
-import javafx.stage.Window;
 import lombok.Getter;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import net.gazeplay.commons.configuration.ActiveConfigurationContext;
 import net.gazeplay.commons.configuration.Configuration;
 import net.gazeplay.commons.ui.Translator;
-import net.gazeplay.commons.utils.CachingSupplier;
 import net.gazeplay.commons.utils.games.BackgroundMusicManager;
 import net.gazeplay.commons.utils.screen.PrimaryScreenDimensionSupplier;
-import net.gazeplay.commons.utils.screen.PrimaryScreenSupplier;
-import net.gazeplay.commons.utils.screen.ScreenDimensionSupplier;
 import net.gazeplay.commons.utils.stats.Stats;
 import net.gazeplay.components.CssUtil;
 import net.gazeplay.gameslocator.GamesLocator;
@@ -34,7 +26,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.stereotype.Component;
 
-import java.util.concurrent.TimeUnit;
 import java.util.function.Supplier;
 
 @Slf4j
@@ -61,8 +52,9 @@ public class GazePlay {
     @Getter
     private PrimaryScreenDimensionSupplier primaryScreenDimensionSupplier;
 
+    @Autowired
     @Getter
-    private Supplier<Dimension2D> currentScreenDimensionSupplier = new CachingSupplier<>(new ScreenDimensionSupplier(new CurrentScreenSupplier(this)), 2, TimeUnit.SECONDS);
+    private Supplier<Dimension2D> currentScreenDimensionSupplier;
 
     @Autowired
     private ApplicationContext applicationContext;
@@ -132,52 +124,4 @@ public class GazePlay {
         return primaryStage.fullScreenProperty();
     }
 
-    @Slf4j
-    @RequiredArgsConstructor
-    private static class CurrentScreenSupplier implements Supplier<Screen> {
-
-        @NonNull
-        private final GazePlay gazePlay;
-
-        private final PrimaryScreenSupplier primaryScreenSupplier = new PrimaryScreenSupplier();
-
-        @Override
-        public Screen get() {
-            log.warn("get()");
-            if (gazePlay.primaryScene == null) {
-                log.warn("primaryScene is null");
-                return primaryScreenSupplier.get();
-            }
-            //double x = gazePlay.primaryScene.getX();
-            //double y = gazePlay.primaryScene.getY();
-
-            double x;
-            double y;
-
-            Window window = gazePlay.primaryScene.getWindow();
-            if (window == null) {
-                log.warn("window is null");
-                Stage primaryStage = gazePlay.getPrimaryStage();
-                if (primaryStage == null) {
-                    log.warn("primaryStage is null");
-                    return primaryScreenSupplier.get();
-                }
-                x = primaryStage.getX() + 1;
-                y = primaryStage.getY() + 1;
-            } else {
-                x = window.getX();
-                y = window.getY();
-            }
-
-            log.info("x = {}, y = {}", x, y);
-            ObservableList<Screen> screensForRectangle = Screen.getScreensForRectangle(x, y, x, y);
-            log.info("screensForRectangle.size() = {}", screensForRectangle.size());
-            if (screensForRectangle.isEmpty()) {
-                return primaryScreenSupplier.get();
-            }
-            Screen screen = screensForRectangle.get(0);
-            log.info("current screen = {}", screen);
-            return screen;
-        }
-    }
 }
